Move schema types into seperate file
